demo/requirements.txt
confluent-kafka==2.5.3
unique-namer==1.6.1